moved "attach hero to unit" to unit rules
