+++ src/bfield/rules/srd/ArmySRD.java
@@ -136,32 +136,4 @@ public class ArmySRD extends bfield.rules.ArmyRules {
     return value  + army.getAttackMod();
   }
 
-  @Override
-  public boolean embedHero(Unit unitToEmbed) {
-    TextInputDialog dialog = new TextInputDialog("8");
-     dialog.setTitle("Attach hero unit");
-     dialog.setHeaderText("Add the encounter level (EL) of the heroes embedded\n"
-             + " on that unit. If the unit is destroyed, the heroes' fate will be\n"
-             + " decided by rules in chapter 6 of the BRCS.");
-     dialog.setContentText("Enter Hero EL:");
-
-     //dialog.getDialogPane().getStylesheets()
-     //  .add(getClass().getResource("dialog.css").toExternalForm());
-     //dialog.getDialogPane().getStyleClass().add("dlg");
-
-     Optional<String> result = dialog.showAndWait();
-     if (result.isPresent()) {
-       try {
-         unitToEmbed.attachHeroUnit(Integer.parseInt(result.get()));
-         return true;
-       } catch (NumberFormatException e) {
-         Alert alert = new Alert (Alert.AlertType.INFORMATION);
-         alert.setTitle("Could not parse number");
-         alert.setHeaderText(null);
-         alert.setContentText(result.get() + " is not a valid integer number.");
-         return false;
-       }
-     } else
-       return false;
-  }
 }
